Final implementation of settings, SOCKS5 proxy, error handling and onboarding
- Cleaned up sidebar by using built-in WinUI Settings button.
- Implemented robust SOCKS5 proxy support with authentication using SocketsHttpHandler.
- Improved API error reporting and fixed URI construction bugs.
- Implemented startup API key check with automatic redirection to Settings.
- Refined UI with InfoBar prompts and fixed data binding types.

+++ ViewModels/SettingsViewModel.cs
@@ -16,7 +16,7 @@ public partial class SettingsViewModel : ObservableObject
     private string _proxyHost = string.Empty;
 
     [ObservableProperty]
-    private int _proxyPort;
+    private double _proxyPort;
 
     [ObservableProperty]
     private string _proxyUsername = string.Empty;
@@ -40,7 +40,7 @@ public partial class SettingsViewModel : ObservableObject
         _settings.ApiKey = ApiKey;
         _settings.ProxyEnabled = ProxyEnabled;
         _settings.ProxyHost = ProxyHost;
-        _settings.ProxyPort = ProxyPort;
+        _settings.ProxyPort = (int)ProxyPort;
         _settings.ProxyUsername = ProxyUsername;
         _settings.ProxyPassword = ProxyPassword;
     }
